Remove all manual uses of v8 serialization
- Seems that sometimes the is a versioning mismatch within the binary
  format causes cluster screen the go grey
