+++ src/main/shell-session/shell-session.ts
@@ -16,11 +16,9 @@ import { UserStore } from "../../common/user-store";
 import * as pty from "node-pty";
 import { appEventBus } from "../../common/app-event-bus/event-bus";
 import logger from "../logger";
-import type { TerminalMessage } from "../../renderer/api/terminal-api";
-import { TerminalChannels } from "../../renderer/api/terminal-api";
-import { deserialize, serialize } from "v8";
 import { stat } from "fs/promises";
 import { getOrInsertWith } from "../../common/utils";
+import { type TerminalMessage, TerminalChannels } from "../../common/terminal/channels";
 
 export class ShellOpenError extends Error {
   constructor(message: string, options?: ErrorOptions) {
@@ -161,7 +159,7 @@ export abstract class ShellSession {
   }
 
   protected send(message: TerminalMessage): void {
-    this.websocket.send(serialize(message));
+    this.websocket.send(JSON.stringify(message));
   }
 
   protected async getCwd(env: Record<string, string | undefined>): Promise<string> {
@@ -232,17 +230,19 @@ export abstract class ShellSession {
     });
 
     this.websocket
-      .on("message", (data: string | Uint8Array): void => {
+      .on("message", (rawData: unknown): void => {
         if (!this.running) {
           return void logger.debug(`[SHELL-SESSION]: received message from ${this.terminalId}, but shellProcess isn't running`);
         }
 
-        if (typeof data === "string") {
-          return void logger.silly(`[SHELL-SESSION]: Received message from ${this.terminalId}`, { data });
+        if (!(rawData instanceof Buffer)) {
+          return void logger.error(`[SHELL-SESSION]: Received message non-buffer message.`, { rawData });
         }
 
+        const data = rawData.toString();
+
         try {
-          const message: TerminalMessage = deserialize(data);
+          const message: TerminalMessage = JSON.parse(data);
 
           switch (message.type) {
             case TerminalChannels.STDIN:
